revert config class split
